api/config.py
from os import getenv, path
from dotenv import load_dotenv

basedir = path.abspath(path.dirname(__file__))
load_dotenv(path.join(basedir, '.env'))

class Config:
    """Stores Flask configuration variables loaded from .env"""

    SECRET_KEY = getenv('SECRET_KEY')
    FLASK_ENV = getenv('FLASK_ENV')

    SQLALCHEMY_DATABASE_URI = getenv('SQLALCHEMY_DATABASE_URI')
    SQLALCHEMY_ECHO = True
    SQLALCHEMY_TRACK_MODIFICATIONS = False
